+++ src/program-types.ts
@@ -195,6 +195,12 @@ export interface InsuranceDepositAccount {
 
 export interface WhitelistInsuranceAccount {
   nonce: number;
+  userKey: PublicKey;
+}
+
+export interface WhitelistDepositAccount {
+  nonce: number;
+  userKey: PublicKey;
 }
 
 export interface SocializedLossAccount {
@@ -204,4 +210,5 @@ export interface SocializedLossAccount {
 
 export interface WhitelistTradingFeesAccount {
   nonce: number;
+  userKey: PublicKey;
 }